requirements.txt
APScheduler
Django
djangorestframework
psycopg2
rq
requests


whitenoise
django
gunicorn
django-heroku

